Better snippet structure
